make cnn_method
